Avoid Path.toFile() throwing UnsupportedOperationException
For ZipPath, toFile() is not possible and throws an exception.

So we should not try to get the File of Paths
but rather keep working with the Path

This closes #160

Keep FileRecord for backwards-compatibility.

Add test for zip-file walking

Also verify that the correct type of object is
registered as 'marker' by the instrumentation

Add slightly more testing to check one unexpected
file-handle leak that is reported

Ignore closing of ZipFile/URLJarFile in tests

Add one more test for Files.createTempFile()